% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.R
\name{getData}
\alias{getData}
\title{Fetch the data from Amazon.com}
\usage{
getData(timeout = 1)
}
\arguments{
\item{timeout}{- (default=1) set the time between requests}
}
\description{
Fetch the data from Amazon.com
}

