% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.R
\name{getIndivItems}
\alias{getIndivItems}
\title{to get individual items fromt he "getItems" function}
\usage{
getIndivItems(items, timeout = 1)
}
\arguments{
\item{items,}{the items to get detailed info about}

\item{timeout,}{the amouunt of time to wait between requests}
}
\description{
to get individual items fromt he "getItems" function
}

