% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.R
\name{getItems}
\alias{getItems}
\title{get items}
\usage{
getItems(url, category, subCategory)
}
\arguments{
\item{url}{the url to fetch the items from}

\item{category}{labelling}

\item{subCategory}{labeling}
}
\description{
get items
}

